Update addresses_controller.rb
Modified to not be able to delete if there is a record in item or trade.

+++ app/controllers/addresses_controller.rb
@@ -1,5 +1,6 @@
 class AddressesController < ApplicationController
   before_action :set_address, only: [:edit, :update, :destroy]
+  before_action :must_not_destroy_child_eixts, only: [:destroy]
 
   def index
     @addresses = Address.where(user_id: current_user.id)
@@ -24,8 +25,12 @@ class AddressesController < ApplicationController
   end
 
   def destroy
-    render :index and return unless @address.destroy
-    redirect_to user_addresses_path, notice:"住所を削除しました" if @address.destroy
+    if must_not_destroy_child_eixts
+      render :index and return unless @address.destroy
+      redirect_to user_addresses_path, notice:"住所を削除しました" 
+    else
+      redirect_to user_addresses_path, notice:"出品または購入に利用した住所は削除できません" 
+    end
   end
 
   
@@ -37,4 +42,9 @@ class AddressesController < ApplicationController
   def set_address
     @address = Address.find(params[:id])
   end
+
+  def must_not_destroy_child_eixts
+    return @address.items.empty? && @address.trades.empty?  
+  end
+
 end
